[IT-3553] bugfix: don't display empty resources table
It's possible for an email to contain an empty table of owned resources
if all owned resources are in accounts also owned by the same email.

Only build a table for owned resources if there are still entries
left after removing owned accounts, and only include an 'accounts'
subkey if account details have been found.

Adjust the test scenario accordingly, and add a description to it.